+++ src/struct.jl
@@ -157,7 +157,7 @@ mutable struct ParametersOrthophoto
 		ucs::Union{String,Matrix{Float64}},
 		BGcolor::Array{Float64,1},
 		pc::Bool,
-		epsg::Union{Nothing,Int64}
+		epsg::Union{Nothing,Integer}
 		)
 
 		# check validity
